restructure metrics functions to metrics crate for optional dependencies
