+++ node/bft/ledger-service/Cargo.toml
@@ -20,7 +20,7 @@ edition = "2021"
 default = [ ]
 ledger = [ "lru", "parking_lot", "rand", "tokio", "tracing" ]
 ledger-write = [ ]
-metrics = ["dep:metrics", "snarkvm/metrics", "rayon"]
+metrics = ["dep:metrics", "snarkvm/metrics"]
 mock = [ "parking_lot", "tracing" ]
 prover = [ ]
 serial = [ ]
